feat: enables EF Core connection retries and adds comprehensive observability and resilience documentation

+++ OrderService.Api/Application/Metrics/OrderMetrics.cs
@@ -7,7 +7,7 @@ public class OrderMetrics
     public const string MeterName = "Ecommerce.OrderService";
     private readonly Meter _meter;
     private readonly Counter<int> _ordersCreatedCounter;
-    private readonly Counter<decimal> _revenueCounter;
+    private readonly Counter<double> _revenueCounter;
     private readonly Counter<int> _orderSuccessCounter;
     private readonly Counter<int> _orderFailedCounter;
 
@@ -15,7 +15,7 @@ public class OrderMetrics
     {
         _meter = meterFactory.Create(MeterName);
         _ordersCreatedCounter = _meter.CreateCounter<int>("orders.created.count", description: "Number of created orders");
-        _revenueCounter = _meter.CreateCounter<decimal>("revenue.total", description: "Total revenue amount from confirmed orders");
+        _revenueCounter = _meter.CreateCounter<double>("revenue.total", description: "Total revenue amount from confirmed orders");
         _orderSuccessCounter = _meter.CreateCounter<int>("orders.success.count", description: "Number of successfully processed orders");
         _orderFailedCounter = _meter.CreateCounter<int>("orders.failed.count", description: "Number of failed orders");
     }
@@ -24,7 +24,7 @@ public class OrderMetrics
     public void OrderSuccess(decimal amount)
     {
         _orderSuccessCounter.Add(1);
-        _revenueCounter.Add(amount);
+        _revenueCounter.Add((double)amount);
     }
     public void OrderFailed() => _orderFailedCounter.Add(1);
 }
